+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.79
 
 - only pull ratings for visible elements, to reduce API calls
+- fixed TMDB ratings not clickable on Prime Video
 
 ## 1.1.78